libasignify: slightly rework asignify_verify_parse_digest
Process the exceptional ASIGNIFY_DIGEST_SIZE up front so that we can
just drop the symbolic entry it has in the digests_size mapping table.

Second, let's just assume that we only have digests we know about in
digests_sizes; type can be valid above it if ANSIFY_DIGEST_MAX was
specified or if the enum gets expanded while this is overlooked, so
reject any entries beyond the highest specified in the table.

The remainder of items in digests_sizes *must* have a non-zero length,
so let's enforce that. Previously, we would truck on through and process
the incoming data without any sanity checking on dlen.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>