libasignify: verify: slightly remodel the digest parser
Pull errors/finalization out of the state machine.  If we reach the
'final' state, we bail out immediately and trigger a parse error if
we somehow didn't reach it at exactly the end.

If we reach the end and didn't finish on the PARSE_START start, then
that is *also* a parse erorr.

Both of these scenarios should be unlikely as the digest is properly
signed at this point.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>