libasignify: add a version script
Some impending changes to be more friendly for FreeBSD's pkg (and other
applications) are going to break ABI for a large chunk of libasignify
by changing filename arguments to FILEs and adding name arguments in
some places.

Prepare in advance by adding a versioning script while we have the
old version in place, so we can at least offer binary compatibility
going forward.